% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{rmean}
\alias{rmean}
\title{Compute the mean of a numeric vector}
\usage{
rmean(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A single numeric value: the mean of \code{x}.
}
\description{
Computes the arithmetic mean of a numeric vector.
}
